# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Essa é a API de Controle de Caixa, que pode ser utilizada junto com a aplicação para controle de caixa de pequenos négocios",
        "version": "1.0.1",
        "title": "Controle de Caixa API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "carleonepsouza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "https://app-loja.azurewebsites.net/v1/api"
        }
    ],
    "host": "https://app-loja.azurewebsites.net/v1/api",
...
    "basePath": "/v1",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/profiles": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "register profile",
                "description": "This route will be responsible for register an profile",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            },
                            "examples": {
                                "profile": {
                                    "value": {
                                        "role": "admin",
                                        "status": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "409": {
                        "description": "Conflit!"
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Return a list of profiles",
                "description": "This route will be responsible for return a list of profiles",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profiles/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Return a specific profile",
                "description": "This route will be responsible for return an profile by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Update a specific profile",
                "description": "This route will be responsible for update an profile by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            },
                            "examples": {
                                "profile": {
                                    "value": {
                                        "role": "atendimento",
                                        "status": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Delete a specific user",
                "description": "This route will be responsible for delete an profile by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Profile": {
                "type": "object",
                "properties": {
                    "role": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}